building with old notifications removed
